requirements.txt
azure-functions
azure-functions-worker
pyodbc
azure-storage-blob
azure-ai-formrecognizer
python-dateutil